accessible/tests/mochitest/tree/test_invalid_img.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
  <title>invalid html img</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>

  <script>
  <![CDATA[
    function doTest() {
      document.getElementsByTagName("img")[0].firstChild.data = "2";

      var accTree = {
        role: ROLE_TEXT,
        children: [ { role: ROLE_TEXT_LEAF } ],
      };
      testAccessibleTree("the_img", accTree);
      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>
</head>
<body>

  <a target="_blank"
     title="use HyperTextAccessible for invalid img"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=852129">
    Mozilla Bug 852129
  </a>

  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <img id="the_img">1</img>
</body>
</html>
